Condensed the data cleaning for the yearly graphs by species
Originally wanted to do the graphing in a function, but I can't figure out how to make it work. There's something funny happening when I try to do the subsetting by species & then graphing in a function. I think it's some issue with using a string as the function argument. There's probably a simple solution, but for now, I'll just create a new graph for each species of interest. There aren't that many of them.