Created a new script just for the species timeseries & updated years
Needed lots of space for the new specifications of the years, where Nov & Dec from the previous year are combined with the other months from the next year to show true timeseries. Also now excluding the very first bit of 2000 and the end of 2019. 

If I get 2020 data, then I can add the rest of 2019. But that will mean that to have all of 2020, I'd also need 2021.....