Added stdev error bars - larger than the mean for some
This is likely because the distribution in total landings for these species (esp. market squid & pelagics) over the time series do not form a normal distribution. So, this error isn't particularly useful, but I does suggest that just using means to look at how the years with the salmon closures differ from the rest is insufficient.